+++ modules/@ergonode/product-batch-actions/README.md
@@ -1,0 +1,16 @@
+
+<p align="center">
+  <a href="https://ergonode.com" rel="noopener noreferrer">
+    <img width="300" src="https://ergonode.com/img/logo-dark.svg" alt="Ergonode logo">
+  </a>
+</p>
+<h3 align="center">Product batch actions module</h3>
+<p align="center">
+  <a href="https://www.npmjs.com/package/@ergonode/product-batch-actions">
+    <img alt="npm" src="https://img.shields.io/npm/v/@ergonode/product-batch-actions">
+  </a>
+  <a href="https://www.npmjs.com/package/@ergonode/product-batch-actions">
+    <img alt="npm" src="https://img.shields.io/npm/l/@ergonode/product-batch-actions">
+  </a>
+</p>
+<br>
