+++ modules/@ergonode/product-batch-actions/src/components/Forms/Fields/AttributeGalleryFormField.vue
