+++ modules/@ergonode/product-batch-actions/src/components/Forms/Fields/AttributeMultiSelectFormField.vue
