+++ modules/@ergonode/product-batch-actions/src/components/Forms/Fields/AttributeSelectFormField.vue
