+++ modules/@ergonode/product-batch-actions/src/components/Modals/ProductBatchActionDetailsModalGrid.vue
@@ -1,0 +1,37 @@
+/*
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE for license details.
+ */
+<template>
+    <ModalGrid
+        :title="$t('@ProductBatchActions.productBatchAction.components.ProductBatchActionDetailsModalGrid.title')"
+        @close="onClose">
+        <template #body>
+            <ProductBatchActionDetailsGrid :item="item" />
+        </template>
+    </ModalGrid>
+</template>
+
+<script>
+import ProductBatchActionDetailsGrid from '@ProductBatchActions/components/Grids/ProductBatchActionDetailsGrid';
+import ModalGrid from '@UI/components/Modal/ModalGrid';
+
+export default {
+    name: 'ProductBatchActionDetailsModalGrid',
+    components: {
+        ModalGrid,
+        ProductBatchActionDetailsGrid,
+    },
+    props: {
+        item: {
+            type: Object,
+            required: true,
+        },
+    },
+    methods: {
+        onClose() {
+            this.$emit('close');
+        },
+    },
+};
+</script>
