+++ modules/@ergonode/product-batch-actions/src/components/SideBars/AttributeSideBarElement.vue
