--- modules/@ergonode/products-batch-actions/README.md
@@ -1,16 +1,0 @@
-
-<p align="center">
-  <a href="https://ergonode.com" rel="noopener noreferrer">
-    <img width="300" src="https://ergonode.com/img/logo-dark.svg" alt="Ergonode logo">
-  </a>
-</p>
-<h3 align="center">Products batch actions module</h3>
-<p align="center">
-  <a href="https://www.npmjs.com/package/@ergonode/products-batch-actions">
-    <img alt="npm" src="https://img.shields.io/npm/v/@ergonode/products-batch-actions">
-  </a>
-  <a href="https://www.npmjs.com/package/@ergonode/products-batch-actions">
-    <img alt="npm" src="https://img.shields.io/npm/l/@ergonode/products-batch-actions">
-  </a>
-</p>
-<br>
