--- modules/@ergonode/products-batch-actions/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@ergonode/products-batch-actions",
-  "version": "1.0.0-rc.1",
-  "author": "Team Ergonode <team@ergonode.com>",
-  "description": "Ergonode products batch actions module",
-  "license": "OSL-3.0",
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/ergonode/frontend.git",
-    "directory": "modules/@ergonode/products-batch-actions"
-  },
-  "homepage": "https://ergonode.com",
-  "bugs": {
-    "url": "https://github.com/ergonode/frontend/issues"
-  },
-  "main": "src/index.js",
-  "files": [
-    "src/*"
-  ],
-  "keywords": [
-    "node",
-    "npm",
-    "vue",
-    "nuxt",
-    "microservice",
-    "micro-service",
-    "micro services",
-    "microfrontend",
-    "micro-frontend",
-    "micro frontend",
-    "ergonode",
-    "ergonode-module"
-  ]
-}
