--- modules/@ergonode/products-batch-actions/src/components/DropZones/RemoveFormFieldDropZone.vue
@@ -1,51 +1,0 @@
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE for license details.
- */
-<template>
-    <FadeTransition>
-        <DropZone
-            v-show="isDropZoneVisible"
-            :hover-background-color="graphiteLightColor"
-            :title="$t('@ProductsBatchActions.productBatchAction.components.RemoveFormFieldDropZone.title')">
-            <template #icon="{ color }">
-                <IconRemoveFilter :fill-color="color" />
-            </template>
-        </DropZone>
-    </FadeTransition>
-</template>
-
-<script>
-import {
-    DRAGGED_ELEMENT,
-} from '@Core/defaults/grid';
-import {
-    GRAPHITE_LIGHT,
-} from '@UI/assets/scss/_js-variables/colors.scss';
-import DropZone from '@UI/components/DropZone/DropZone';
-import IconRemoveFilter from '@UI/components/Icons/Actions/IconRemoveFilter';
-import FadeTransition from '@UI/components/Transitions/FadeTransition';
-import {
-    mapState,
-} from 'vuex';
-
-export default {
-    name: 'RemoveFormFieldDropZone',
-    components: {
-        IconRemoveFilter,
-        DropZone,
-        FadeTransition,
-    },
-    computed: {
-        ...mapState('draggable', [
-            'isElementDragging',
-        ]),
-        graphiteLightColor() {
-            return GRAPHITE_LIGHT;
-        },
-        isDropZoneVisible() {
-            return this.isElementDragging === DRAGGED_ELEMENT.FORM_FIELD;
-        },
-    },
-};
-</script>
