--- modules/@ergonode/products-batch-actions/src/components/Grids/ProductBatchActionDetailsGrid.vue
@@ -1,177 +1,0 @@
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE for license details.
- */
-<template>
-    <Grid
-        :columns="columns"
-        :data-count="filtered"
-        :rows="rows"
-        :pagination="pagination"
-        :filters="filterValues"
-        :extended-components="extendedGridComponents"
-        :is-prefetching-data="isPrefetchingData"
-        :is-header-visible="true"
-        :is-basic-filter="true"
-        @pagination="onPaginationChange"
-        @sort-column="onColumnSortChange"
-        @filter="onFilterChange"
-        @remove-all-filters="onRemoveAllFilters"
-        v-bind="extendedProps['grid']">
-        <template #actionsHeader>
-            <div class="notification-details-tiles">
-                <Tile
-                    v-for="(detail, index) in details"
-                    :key="index"
-                    :label="detail.label"
-                    :value="detail.value" />
-            </div>
-        </template>
-    </Grid>
-</template>
-
-<script>
-import {
-    ALERT_TYPE,
-} from '@Core/defaults/alerts';
-import {
-    DEFAULT_GRID_FETCH_PARAMS,
-    DEFAULT_GRID_PAGINATION,
-} from '@Core/defaults/grid';
-import extendPropsMixin from '@Core/mixins/extend/extendProps';
-import extendedGridComponentsMixin from '@Core/mixins/grid/extendedGridComponentsMixin';
-import {
-    getParsedFilters,
-} from '@Core/models/mappers/gridDataMapper';
-import {
-    getGridData,
-} from '@Core/services/grid/getGridData.service';
-import Grid from '@UI/components/Grid/Grid';
-import Tile from '@UI/components/Tile/Tile';
-
-export default {
-    name: 'ProductBatchActionDetailsGrid',
-    components: {
-        Tile,
-        Grid,
-    },
-    mixins: [
-        extendPropsMixin({
-            extendedKey: '@ProductsBatchActions/components/Grids/ProductBatchActionDetailsGrid/props',
-            extendedNames: [
-                'grid',
-            ],
-        }),
-        extendedGridComponentsMixin,
-    ],
-    props: {
-        item: {
-            type: Object,
-            required: true,
-        },
-    },
-    async fetch() {
-        await this.onFetchData();
-    },
-    data() {
-        return {
-            columns: [],
-            rows: [],
-            filtered: 0,
-            filterValues: {},
-            isPrefetchingData: true,
-            localParams: DEFAULT_GRID_FETCH_PARAMS(),
-            pagination: DEFAULT_GRID_PAGINATION(),
-        };
-    },
-    computed: {
-        details() {
-            return [
-                {
-                    label: this.$t('@ProductsBatchActions.productBatchAction.components.ProductBatchActionDetailsGrid.detailsLabel'),
-                    value: this.item.createdAt,
-                },
-            ];
-        },
-    },
-    methods: {
-        onPaginationChange(pagination) {
-            this.pagination = pagination;
-            this.localParams.limit = pagination.itemsPerPage;
-            this.localParams.offset = (pagination.page - 1) * pagination.itemsPerPage;
-
-            this.onFetchData();
-        },
-        onRemoveAllFilters() {
-            this.filterValues = {};
-            this.pagination.page = 1;
-            this.localParams.filter = '';
-            this.localParams.offset = 0;
-
-            this.onFetchData();
-        },
-        onFilterChange(filters) {
-            this.filterValues = filters;
-            this.pagination.page = 1;
-            this.localParams.filter = getParsedFilters(filters);
-            this.localParams.offset = (this.pagination.page - 1) * this.pagination.itemsPerPage;
-
-            this.onFetchData();
-        },
-        onColumnSortChange(sortOrder) {
-            this.localParams.sortOrder = sortOrder;
-
-            this.onFetchData();
-        },
-        async onFetchData() {
-            const {
-                sortOrder = {}, ...rest
-            } = this.localParams;
-
-            const params = {
-                ...rest,
-                ...sortOrder,
-            };
-
-            await getGridData({
-                $route: this.$route,
-                $cookies: this.$cookies,
-                $axios: this.$axios,
-                path: `batch-action/${this.item.objectId}/entries`,
-                params,
-                onSuccess: this.onFetchGridDataSuccess,
-                onError: this.onFetchGridDataError,
-            });
-        },
-        onFetchGridDataError() {
-            this.$addAlert({
-                type: ALERT_TYPE.ERROR,
-                message: this.$t('@ProductsBatchActions.productBatchAction.components.ProductBatchActionDetailsGrid.getRequest'),
-            });
-
-            this.isPrefetchingData = false;
-        },
-        onFetchGridDataSuccess({
-            columns,
-            rows,
-            filtered,
-        }) {
-            this.columns = columns;
-            this.rows = rows;
-            this.filtered = filtered;
-            this.isPrefetchingData = false;
-        },
-    },
-};
-</script>
-
-<style lang="scss" scoped>
-.notification-details-tiles {
-    display: flex;
-    flex-wrap: wrap;
-
-    & > * {
-        margin-right: 12px;
-    }
-}
-</style>
