--- modules/@ergonode/products-batch-actions/src/locales/en_GB.json
@@ -1,56 +1,0 @@
-{
-  "@ProductsBatchActions": {
-    "productBatchAction": {
-      "_": {
-        "title": "Product attributes",
-        "getRequest": "Attribute hasn’t been fetched properly",
-        "submit": "SAVE CHANGES"
-      },
-      "components": {
-        "UpdateProductsButton": {
-          "confirmTitleSingular": "Are you sure you want to clear the value of 1 attribute?",
-          "confirmTitlePlural": "Are you sure you want to clear the values of {info} attributes?",
-          "confirmSubtitle": "Saving an empty attribute will clear its value in all edited products.",
-          "confirmApplySingular": "Clear 1 attribute",
-          "confirmApplyPlural": "Clear {info} attributes",
-          "emptyAttributes": "No selected attributes",
-          "confirmCountTitle": "Processing of {count} products",
-          "confirmCountSubtitle": "Are you sure you want to run a batch action for {count} products?",
-          "confirmCountApplyPlural": "Process {count} products",
-          "confirmCancel": "Cancel"
-        },
-        "UpdateProductsModal": {
-          "title": "Editing {info} products",
-          "getComponent": "There is not defined attribute component for {info} type.",
-          "draggableFormTitle": "Edited attributes"
-        },
-        "ProductsBatchActions": {
-          "deleteBatchActionLabel": "Delete products",
-          "editBatchActionLabel": "Edit products",
-          "confirmTitle": "Are you sure you want to permanently delete {info} products?",
-          "confirmSubtitle": "The products will be deleted from the system forever and cannot be restored.",
-          "confirmApplyTitle": "DELETE {info} PRODUCTS",
-          "successAlert": "Products have been removed"
-        },
-        "RemoveFormFieldDropZone": {
-          "title": "Remove item"
-        },
-        "AttributesSideBar": {
-          "noAttributesTitle": "No attributes",
-          "noAttributesSubtitle": "There are no common attributes for selected products.",
-          "notAssigned": "Not assigned"
-        },
-        "AttributeNotDefinedField": {
-          "title": "Attribute for {info} type is not defined."
-        },
-        "ProductBatchActionDetailsGrid": {
-          "detailsLabel": "Date of start",
-          "getRequest": "Products batch action details haven’t been fetched properly"
-        },
-        "ProductBatchActionDetailsModalGrid": {
-          "title": "Details"
-        }
-      }
-    }
-  }
-}
