--- modules/@ergonode/products-batch-actions/src/locales/pl_PL.json
@@ -1,56 +1,0 @@
-{
-  "@ProductsBatchActions": {
-    "productBatchAction": {
-      "_": {
-        "title": "Atrybuty produktów",
-        "getRequest": "Atrybut nie został pobrany prawidłowo",
-        "submit": "ZAPISZ ZMIANY"
-      },
-      "components": {
-        "UpdateProductsButton": {
-          "confirmTitleSingular": "Czy jesteś pewien, że chcesz wyczyścić wartość dla 1 atrybutu?",
-          "confirmTitlePlural": "Czy jesteś pewien, że chcesz wyczyścić wartości dla {info} atrybutów?",
-          "confirmSubtitle": "Zapisywanie pustego atrybutu spowoduje wyczyszczenie jego wartości dla wszystkich edytowanych produktów.",
-          "confirmApplySingular": "Wyczyść 1 atrybut",
-          "confirmApplyPlural": "Wyczyść {info} atrybuty",
-          "emptyAttributes": "Brak wybranych atrybuty",
-          "confirmCountTitle": "Przetwarzanie {count} produktów",
-          "confirmCountSubtitle": "Czy jesteś pewny że chcesz uruchomić akcję masową dla {count} produktów?",
-          "confirmCountApplyPlural": "Przetwórz {count} poroduktów",
-          "confirmCancel": "Anuluj"
-        },
-        "UpdateProductsModal": {
-          "title": "Edytowanie {info} produktów",
-          "getComponent": "Nie ma zdefiniowanego komponentu atrybutu dla typu {info}.",
-          "draggableFormTitle": "Edytowane atrybuty"
-        },
-        "ProductsBatchActions": {
-          "deleteBatchActionLabel": "Usuń produkty",
-          "editBatchActionLabel": "Edytuj produkty",
-          "confirmTitle": "Czy jesteś pewień usunięcia permamentnie {info} produktów?",
-          "confirmSubtitle": "Produkty będą usunięte z systemu bez odwrotnie i nie będą mogłby być przywrócone.",
-          "confirmApplyTitle": "USUŃ {info} PRODUKTY",
-          "successAlert": "Produkty zostały usunięte"
-        },
-        "RemoveFormFieldDropZone": {
-          "title": "Usuń element"
-        },
-        "AttributesSideBar": {
-          "noAttributesTitle": "Brak atrybutów",
-          "noAttributesSubtitle": "Brak wspólnych atrybutów dla wybranych produktów.",
-          "notAssigned": "Nie przypisane"
-        },
-        "AttributeNotDefinedField": {
-          "title": "Atrybut dla {info} typu jest nie zdefiniowany."
-        },
-        "ProductBatchActionDetailsGrid": {
-          "detailsLabel": "Data startu",
-          "getRequest": "Szczegóły akcji masowej produktów nie zostały pobrane poprawnie"
-        },
-        "ProductBatchActionDetailsModalGrid": {
-          "title": "Szczegóły"
-        }
-      }
-    }
-  }
-}
